Move colleagues river data to sample_data, move context to context utils, update texts and format
